persmission crud ..

+++ app/Http/Controllers/TaskController.php
@@ -17,11 +17,15 @@ class TaskController extends Controller
 
     public function create()
     {
+        // $this->authorize('manage tasks');
+
         return view('tasks.create');
     }
 
     public function store(StoreTaskRequest $request)
     {
+        $this->authorize('manage tasks');
+
         Task::create($request->validated());
 
         return redirect()->route('tasks.index');
@@ -29,11 +33,15 @@ class TaskController extends Controller
 
     public function edit(Task $task)
     {
+        $this->authorize('manage tasks');
+
         return view('tasks.edit', compact('task'));
     }
 
     public function update(UpdateTaskRequest $request, Task $task)
     {
+        $this->authorize('manage tasks');
+
         $task->update($request->validated());
 
         return redirect()->route('tasks.index');
@@ -41,6 +49,8 @@ class TaskController extends Controller
 
     public function destroy(Task $task)
     {
+        $this->authorize('manage tasks');
+
         $task->delete();
 
         return redirect()->route('tasks.index');
